Lifting sort of working
Flags are more or less functional but logic is not completely correct,
branching seems to be wrong, and make no effort to correctly handle
operand width. This will likely require a rewrite but at least we have a
semi-functional point that we can compare against and figured out how
most of the types work together.