Return flag write types (#4)
This was originally commented out due to a bug in Binary Ninja that
caused the python scripting provider to break when it attempted to
lookup invalid flag classes. The rust bindings were incorrectly
returning the flag write type id rather than the flag class id which
would always have a non-zero value when flag writes exist even if a flag
class does not.